requirements.txt
amqp==5.1.1
asgiref==3.5.2
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.6
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.0.4
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
gunicorn==20.1.0
joblib==1.1.0
kombu==5.2.4
nltk==3.7
numpy==1.22.4
pandas==1.3.1
Pillow==9.1.1
prompt-toolkit==3.0.29
psycopg2-binary==2.9.3
PyJWT==1.7.1
python-dateutil==2.8.2
pytz==2022.1
regex==2022.4.24
scikit-learn==1.1.1
scipy==1.8.1
six==1.16.0
sklearn==0.0
sqlparse==0.4.2
threadpoolctl==3.1.0
tqdm==4.64.0
vine==5.0.0
wcwidth==0.2.5
